cleanup: main städad
